chore: migrated types and fixed incompatible struct pointers
