+++ common/types/inventory/methods.go
@@ -2,24 +2,9 @@ package types
 
 import (
 	"handworks/common/grpc/inventory"
-	"time"
 
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
-
-type DbInventoryItem struct {
-	ID          string
-	Name        string
-	Type        string // stored as string in DB (maps to enum ItemType)
-	Status      string // stored as string in DB (maps to enum ItemStatus)
-	Category    string // stored as string in DB (maps to enum ItemCategory)
-	Quantity    int32
-	MaxQuantity int32
-	Unit        string
-	IsAvailable bool
-	CreatedAt   time.Time
-	UpdatedAt   time.Time
-}
 
 func (dbItem DbInventoryItem) ToProto() *inventory.InventoryItem {
 	return &inventory.InventoryItem{
@@ -30,6 +15,7 @@ func (dbItem DbInventoryItem) ToProto() *inventory.InventoryItem {
 		Category:    inventory.ItemCategory(inventory.ItemCategory_value[dbItem.Category]),
 		Quantity:    dbItem.Quantity,
 		MaxQuantity: dbItem.MaxQuantity,
+		ImageUrl:    dbItem.ImageUrl,
 		Unit:        dbItem.Unit,
 		IsAvailable: dbItem.IsAvailable,
 		CreatedAt:   timestamppb.New(dbItem.CreatedAt),
