+++ docker-compose.yml
@@ -1,1 +1,8 @@
-#TODO: generic docker compose file for all services and nats
+services:
+  nats:
+    image: nats:latest
+    container_name: nats
+    ports:
+      - "4222:4222" # client port
+      - "8222:8222" # monitoring / admin
+    command: ["-js"]
