+++ services/inventory/server/server.go
@@ -1,19 +1,18 @@
 package server
 
 import (
+	"context"
 	"fmt"
 	"handworks-services-inventory/service"
 	"handworks/common/grpc/inventory"
 	"handworks/common/utils"
 	"net"
 	"os"
-	"os/signal"
-	"syscall"
 
 	"google.golang.org/grpc"
 )
 
-func StartGrpcServer(s *service.InventoryService, logger *utils.Logger) error {
+func StartGrpcServer(ctx context.Context, s *service.InventoryService, logger *utils.Logger) error {
 	host := getEnv("DEV_URL", "localhost")
 	port := getEnv("PORT", "9090")
 	addr := fmt.Sprintf("%s:%s", host, port)
@@ -26,17 +25,14 @@ func StartGrpcServer(s *service.InventoryService, logger *utils.Logger) error {
 	grpcServer := grpc.NewServer()
 	inventory.RegisterInventoryServiceServer(grpcServer, s)
 
-	// graceful shutdown
 	go func() {
-		c := make(chan os.Signal, 1)
-		signal.Notify(c, os.Interrupt, syscall.SIGTERM)
-		<-c
-		logger.Warn("Shutting down Inventory gRPC server...")
+		<-ctx.Done()
+		logger.Warn("Gracefully stopping gRPC server...")
 		grpcServer.GracefulStop()
 		lis.Close()
 	}()
 
-	logger.Info("Inventory Service gRPC server listening on %s", addr)
+	logger.Info("Inventory gRPC server listening on %s", addr)
 	return grpcServer.Serve(lis)
 }
